requirements.txt
Flask==2.3.2
Flask-PyMongo==2.3.0
CustomTkinter
dnspython